Remove old files and commit reformatting by prettier
