+++ src/Omise.js
@@ -4,110 +4,104 @@
  * -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=
  */
 
-import { isUri } from 'valid-url';
-import 'whatwg-fetch';
+import { isUri } from 'valid-url'
+import 'whatwg-fetch'
 
 export default function OmiseFactory(config) {
+  const result = verifyConfigStructure(config)
+  if (result.error) throw new Error(result.message)
 
-  const result = verifyConfigStructure(config);
-  if (result.error) throw new Error(result.message);
+  const RPC_TIMEOUT = 30000
 
-  const RPC_TIMEOUT = 30000;
-
-  let
-
-  //----------------- Public interface
+  let //----------------- Public interface
 
     Omise = {
       config,
       publicKey: null,
       setPublicKey,
       createSource,
-      createToken
+      createToken,
     },
-
-  // ------------------ Private vars
+    // ------------------ Private vars
 
     _rpc = null
 
-  ;
-
-
   function _createRpc(callback) {
     if (_rpc) {
-      return _rpc;
+      return _rpc
     } else {
-      const { vaultUrl } = Omise.config;
+      const { vaultUrl } = Omise.config
       const tm = setTimeout(() => {
         _rpc.destroy()
         _rpc = null
         callback && callback()
-      }, RPC_TIMEOUT);
+      }, RPC_TIMEOUT)
 
-      _rpc = new easyXDM.Rpc({
-        remote: `${vaultUrl}/provider`,
-        onReady() {
-          clearTimeout(tm)
-        }
-      }, {
-        remote: {
-          createToken: {}
+      _rpc = new easyXDM.Rpc(
+        {
+          remote: `${vaultUrl}/provider`,
+          onReady() {
+            clearTimeout(tm)
+          },
         },
-      });
+        {
+          remote: {
+            createToken: {},
+          },
+        }
+      )
 
-      return _rpc;
+      return _rpc
     }
   }
 
   function setPublicKey(publicKey) {
-    Omise.publicKey = publicKey;
-    return Omise.publicKey;
+    Omise.publicKey = publicKey
+    return Omise.publicKey
   }
 
   function createSource(type, options, callback) {
-    const auth = btoa(Omise.publicKey);
+    const auth = btoa(Omise.publicKey)
 
-    options.type = type;
+    options.type = type
 
-    const url = `${Omise.config.interfaceUrl}/sources/`;
+    const url = `${Omise.config.interfaceUrl}/sources/`
 
     fetch(url, {
-
       method: 'post',
       headers: {
         Authorization: `Basic ${auth}`,
         'Content-Type': 'application/json',
       },
-      body: JSON.stringify(options)
-
-    }).then(response =>
-      response.json().then(data => callback(response.status, data))
-
-    ).catch(error => {
-
-      callback(0, {
-        code: 'create_source_error',
-        error: error.message
+      body: JSON.stringify(options),
+    })
+      .then(response =>
+        response.json().then(data => callback(response.status, data))
+      )
+      .catch(error => {
+        callback(0, {
+          code: 'create_source_error',
+          error: error.message,
+        })
       })
-    });
   }
 
   function createToken(as, attributes, handler) {
-    const data = { [as]: attributes };
+    const data = { [as]: attributes }
     _createRpc(() => {
       handler(0, {
         code: 'rpc_error',
-        message: 'unable to connect to provider after timeout'
+        message: 'unable to connect to provider after timeout',
       })
     }).createToken(
       Omise.publicKey,
       data,
       response => handler(response.status, response.data),
       e => handler(e.data.status, e.data.data)
-    );
+    )
   }
 
-  return Omise;
+  return Omise
 }
 
 /**
@@ -117,18 +111,18 @@ export default function OmiseFactory(config) {
 export function verifyConfigStructure(config) {
   const result = {
     error: false,
-    message: ''
-  };
-
-  if (!config.vaultUrl || !isUri(config.vaultUrl)) {
-    result.message = 'Missing vaultUrl';
-  } else if (!config.cardHost || !isUri(config.cardHost)) {
-    result.message = 'Missing cardHost';
-  } else if (!config.interfaceUrl || !isUri(config.interfaceUrl)) {
-    result.message = 'Missing interfaceUrl';
+    message: '',
   }
 
-  result.error = result.message !== '';
+  if (!config.vaultUrl || !isUri(config.vaultUrl)) {
+    result.message = 'Missing vaultUrl'
+  } else if (!config.cardHost || !isUri(config.cardHost)) {
+    result.message = 'Missing cardHost'
+  } else if (!config.interfaceUrl || !isUri(config.interfaceUrl)) {
+    result.message = 'Missing interfaceUrl'
+  }
 
-  return result;
+  result.error = result.message !== ''
+
+  return result
 }
