+++ src/index.js
@@ -4,20 +4,19 @@
  * -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=
  */
 
-import 'helpers/polyfill';
+import 'helpers/polyfill'
 
-import 'vendors/easyXDM.js';
+import 'vendors/easyXDM.js'
 
-import config from './config';
-import OmiseFactory from './Omise';
-import OmiseCardFactory from './OmiseCard';
+import config from './config'
+import OmiseFactory from './Omise'
+import OmiseCardFactory from './OmiseCard'
 
-export default OmiseFactory;
+export default OmiseFactory
 
 /* dev:start */
-console.log('Development mode');
+console.log('Development mode')
 /* dev:end */
 
-global.Omise = OmiseFactory(config);
-global.OmiseCard = OmiseCardFactory(config);
-
+global.Omise = OmiseFactory(config)
+global.OmiseCard = OmiseCardFactory(config)
